guardrails/pyproject.toml
[project]
name = "guardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "detoxify>=0.5.2",
    "python-dotenv>=1.0.1",
    "torch>=2.6.0",
    "nltk>=3.8.1",
    "presidio-analyzer>=2.2.357",
    "presidio-anonymizer>=2.2.357",
    "transformers>=4.48.3",
    "guardrails-ai>=0.6.3",
    "guardrails-api-client>=0.4.0a1",
    "typer>=0.15.2"
]

